fix for download and upload

+++ core/src/androidMain/kotlin/com/ustadmobile/core/catalog/contenttype/VideoTypePluginAndroid.kt
@@ -244,7 +244,9 @@ class VideoTypePluginAndroid(
                     mediaTransformer.release()
                     newVideo.delete()
                     videoTempDir.deleteRecursively()
-                    localUri.toFile().delete()
+                    if(videoUri.isRemote()){
+                        localUri.toFile().delete()
+                    }
                 }
 
                 throw c
